Refactor client dashbaord into several files
* client dashbaord is now several files instead of a 1.4k line monster
* remove the wifi settings endpoint as it's no longer viable (legacy features) and we don't use it
* change out the wifi ssid and password endpoitns not to nest structs pointlessly
