+++ rita/src/rita_client/dashboard/wifi.rs
@@ -1,0 +1,95 @@
+/*
+These endpoints are used to modify mundane wireless settings
+*/
+
+use actix::prelude::*;
+use failure::Error;
+
+use rita_common::dashboard::Dashboard;
+use KI;
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct WifiInterface {
+    #[serde(default)]
+    pub section_name: String,
+    pub network: String,
+    #[serde(default)]
+    pub mesh: bool,
+    pub mode: String,
+    pub ssid: String,
+    pub encryption: String,
+    pub key: String,
+    #[serde(default, skip_deserializing)]
+    pub device: WifiDevice,
+}
+
+#[derive(Serialize, Deserialize, Default, Clone, Debug)]
+pub struct WifiDevice {
+    #[serde(default)]
+    pub section_name: String,
+    #[serde(rename = "type")]
+    pub i_type: String,
+    pub channel: String,
+    pub path: String,
+    pub htmode: String,
+    pub hwmode: String,
+    pub disabled: String,
+    #[serde(default)]
+    pub radio_type: String,
+}
+
+#[derive(Serialize, Deserialize, Default, Clone, Debug)]
+pub struct WifiSSID {
+    pub radio: String,
+    pub ssid: String,
+}
+
+#[derive(Serialize, Deserialize, Default, Clone, Debug)]
+pub struct WifiPass {
+    pub radio: String,
+    pub pass: String,
+}
+
+impl Message for WifiSSID {
+    type Result = Result<(), Error>;
+}
+
+impl Handler<WifiSSID> for Dashboard {
+    type Result = Result<(), Error>;
+    fn handle(&mut self, msg: WifiSSID, _ctx: &mut Self::Context) -> Self::Result {
+        // think radio0, radio1
+        let iface_name = msg.radio;
+        let ssid = msg.ssid;
+        let section_name = format!("default_{}", iface_name);
+        KI.set_uci_var(&format!("wireless.{}.ssid", section_name), &ssid)?;
+
+        KI.uci_commit()?;
+        KI.openwrt_reset_wireless()?;
+
+        // We edited disk contents, force global sync
+        KI.fs_sync()?;
+        Ok(())
+    }
+}
+
+impl Message for WifiPass {
+    type Result = Result<(), Error>;
+}
+
+impl Handler<WifiPass> for Dashboard {
+    type Result = Result<(), Error>;
+    fn handle(&mut self, msg: WifiPass, _ctx: &mut Self::Context) -> Self::Result {
+        // think radio0, radio1
+        let iface_name = msg.radio;
+        let pass = msg.pass;
+        let section_name = format!("default_{}", iface_name);
+        KI.set_uci_var(&format!("wireless.{}.key", section_name), &pass)?;
+
+        KI.uci_commit()?;
+        KI.openwrt_reset_wireless()?;
+
+        // We edited disk contents, force global sync
+        KI.fs_sync()?;
+        Ok(())
+    }
+}
